requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.11.0
openai==2.0.0
tiktoken==0.7.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
pytest-mock==3.15.1
python-dotenv==1.1.1
PyYAML==6.0.3
setuptools==80.9.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
wheel==0.45.1
